requirements.txt
discord.py
python-dotenv
psutil
humanize
requests
aiohttp